+++ web/src/features/landingPage/components/Sidebar.tsx
@@ -13,10 +13,10 @@ import Typography from '@mui/material/Typography'
 import useMediaQuery from '@mui/material/useMediaQuery'
 import makeStyles from '@mui/styles/makeStyles'
 import { LANDING_BACKGROUND_COLOUR } from 'app/theme'
-import SidebarToolList from 'landingPage/SidebarToolList'
-import Subheading from 'landingPage/Subheading'
-import { ReactComponent as MsTeamsIcon } from 'landingPage/msTeams.svg'
-import { ReactComponent as MiroIcon } from 'landingPage/miro.svg'
+import SidebarToolList from 'features/landingPage/components/SidebarToolList'
+import Subheading from 'features/landingPage/components/Subheading'
+import { ReactComponent as MsTeamsIcon } from 'features/landingPage/images/msTeams.svg'
+import { ReactComponent as MiroIcon } from 'features/landingPage/images/miro.svg'
 
 const msTeamsSprintMeeting =
   'https://teams.microsoft.com/l/meetup-join/19%3ameeting_NDc0MDcwZTMtNTMxMS00MzEwLTg3MmItZGFiMjY2YjRjZWU1%40thread.v2/0?context=%7b%22Tid%22%3a%226fdb5200-3d0d-4a8a-b036-d3685e359adc%22%2c%22Oid%22%3a%22621f41d9-25a1-447c-9edd-9accf5794ad6%22%7d'
